backend/requirements.txt
Django==5.1.1
djangorestframework==3.15.2
pandas==2.2.2
openpyxl==3.1.5
django-cors-headers==4.4.0

